pendulum swings

+++ pendulum/lcd.h
@@ -1,14 +1,14 @@
 #ifndef LeastCommonMultiple_h
 #define LeastCommonMultiple_h
 
-int gcd(int a, int b) {  
+unsigned long gcd(int a, int b) {  
   if (a == 0) 
     return b;  
   return gcd(b % a, a);  
 }  
   
 // Function to return LCM of two numbers  
-int lcm(int a, int b) {  
+unsigned long lcm(unsigned long a, unsigned long b) {  
   return (a*b)/gcd(a, b);  
 }
